Refactoring - fixed wrong imports
