+++ nostr-event/src/main/java/nostr/event/impl/DirectMessageEvent.java
@@ -2,8 +2,8 @@ package nostr.event.impl;
 
 import nostr.base.PublicKey;
 import nostr.event.Kind;
-import nostr.base.list.TagList;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 
 /**
  *
