+++ nostr-event/src/main/java/nostr/event/impl/GenericEvent.java
@@ -24,10 +24,10 @@ import nostr.base.PublicKey;
 import nostr.base.Signature;
 import nostr.base.annotation.JsonString;
 import nostr.base.annotation.Key;
-import nostr.base.list.TagList;
 import nostr.crypto.bech32.Bech32;
 import nostr.event.BaseEvent;
 import nostr.event.Kind;
+import nostr.event.list.TagList;
 import nostr.event.marshaller.impl.EventMarshaller;
 import nostr.event.marshaller.impl.TagListMarshaller;
 import nostr.util.NostrException;
