+++ nostr-event/src/main/java/nostr/event/impl/InternetIdentifierMetadataEvent.java
@@ -3,7 +3,6 @@ package nostr.event.impl;
 import nostr.base.Profile;
 import nostr.event.Kind;
 import nostr.base.PublicKey;
-import nostr.base.list.TagList;
 import java.beans.IntrospectionException;
 import java.lang.reflect.InvocationTargetException;
 import java.security.NoSuchAlgorithmException;
@@ -15,6 +14,7 @@ import lombok.EqualsAndHashCode;
 import lombok.NonNull;
 import lombok.extern.java.Log;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 import nostr.event.util.Nip05Validator;
 import nostr.types.values.IValue;
 import nostr.types.values.impl.ExpressionValue;
