+++ nostr-event/src/main/java/nostr/event/impl/MetadataEvent.java
@@ -4,7 +4,6 @@ package nostr.event.impl;
 import nostr.base.Profile;
 import nostr.event.Kind;
 import nostr.base.PublicKey;
-import nostr.base.list.TagList;
 import java.beans.IntrospectionException;
 import java.lang.reflect.InvocationTargetException;
 import java.security.NoSuchAlgorithmException;
@@ -15,6 +14,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.extern.java.Log;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 import nostr.types.values.IValue;
 import nostr.types.values.impl.ExpressionValue;
 import nostr.types.values.impl.ObjectValue;
