+++ nostr-event/src/main/java/nostr/event/impl/ReactionEvent.java
@@ -3,11 +3,11 @@ package nostr.event.impl;
 import nostr.event.Kind;
 import nostr.base.PublicKey;
 import nostr.event.Reaction;
-import nostr.base.list.TagList;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.extern.java.Log;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 
 /**
  *
