+++ nostr-event/src/main/java/nostr/event/impl/ReplaceableEvent.java
@@ -2,11 +2,11 @@ package nostr.event.impl;
 
 import nostr.event.Kind;
 import nostr.base.PublicKey;
-import nostr.base.list.TagList;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.extern.java.Log;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 
 /**
  *
