+++ nostr-event/src/main/java/nostr/event/impl/TextNoteEvent.java
@@ -3,8 +3,8 @@ package nostr.event.impl;
 
 import nostr.base.PublicKey;
 import nostr.event.Kind;
-import nostr.base.list.TagList;
 import nostr.base.annotation.Event;
+import nostr.event.list.TagList;
 
 /**
  *
