+++ nostr-event/src/main/java/nostr/event/list/EventList.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.NonNull;
 import lombok.extern.java.Log;
 import nostr.base.annotation.JsonList;
-import nostr.base.list.BaseList;
 import nostr.event.impl.GenericEvent;
 
 /**
