+++ nostr-event/src/main/java/nostr/event/list/GenericTagQueryList.java
@@ -1,4 +1,4 @@
-package nostr.base.list;
+package nostr.event.list;
 
 import java.util.ArrayList;
 import java.util.List;
