+++ nostr-event/src/main/java/nostr/event/list/KindList.java
@@ -1,5 +1,5 @@
 
-package nostr.base.list;
+package nostr.event.list;
 
 import nostr.event.Kind;
 import java.util.ArrayList;
