+++ nostr-event/src/main/java/nostr/event/marshaller/impl/TagListMarshaller.java
@@ -3,12 +3,12 @@ package nostr.event.marshaller.impl;
 import nostr.base.ITag;
 import nostr.base.Relay;
 import nostr.event.marshaller.BaseListMarhsaller;
-import nostr.base.list.TagList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 import lombok.extern.java.Log;
 import static nostr.base.NipUtil.checkSupport;
+import nostr.event.list.TagList;
 import nostr.util.NostrException;
 
 /**
