+++ nostr-event/src/main/java/nostr/event/unmarshaller/impl/FiltersUnmarshaller.java
@@ -10,10 +10,10 @@ import nostr.base.PublicKey;
 import nostr.event.BaseEvent.ProxyEvent;
 import nostr.event.Kind;
 import nostr.event.impl.Filters;
-import nostr.base.list.GenericTagQueryList;
-import nostr.base.list.KindList;
-import nostr.base.list.PublicKeyList;
 import nostr.event.list.EventList;
+import nostr.event.list.GenericTagQueryList;
+import nostr.event.list.KindList;
+import nostr.event.list.PublicKeyList;
 import nostr.event.unmarshaller.BaseElementUnmarshaller;
 import nostr.json.unmarshaller.impl.JsonObjectUnmarshaller;
 import nostr.types.values.impl.ArrayValue;
