+++ nostr-test/src/main/java/nostr/test/EntityFactory.java
@@ -14,10 +14,6 @@ import nostr.event.tag.PubKeyTag;
 import nostr.event.impl.ReactionEvent;
 import nostr.event.impl.ReplaceableEvent;
 import nostr.event.impl.TextNoteEvent;
-import nostr.base.list.KindList;
-import nostr.base.list.PubKeyTagList;
-import nostr.base.list.PublicKeyList;
-import nostr.base.list.TagList;
 import java.beans.IntrospectionException;
 import java.lang.reflect.InvocationTargetException;
 import java.net.MalformedURLException;
@@ -32,8 +28,12 @@ import nostr.base.GenericTagQuery;
 import nostr.base.IEvent;
 import nostr.event.impl.GenericTag;
 import nostr.event.impl.OtsEvent;
-import nostr.base.list.GenericTagQueryList;
 import nostr.event.list.EventList;
+import nostr.event.list.GenericTagQueryList;
+import nostr.event.list.KindList;
+import nostr.event.list.PubKeyTagList;
+import nostr.event.list.PublicKeyList;
+import nostr.event.list.TagList;
 import nostr.types.values.impl.StringValue;
 import nostr.util.NostrException;
