+++ nostr-test/src/test/java/nostr/test/event/EventTest.java
@@ -17,11 +17,11 @@ import nostr.event.impl.Filters;
 import nostr.event.impl.GenericMessage;
 import nostr.event.impl.GenericTag;
 import nostr.event.impl.TextNoteEvent;
-import nostr.base.list.GenericTagQueryList;
-import nostr.base.list.KindList;
-import nostr.base.list.PublicKeyList;
-import nostr.base.list.TagList;
 import nostr.event.list.EventList;
+import nostr.event.list.GenericTagQueryList;
+import nostr.event.list.KindList;
+import nostr.event.list.PublicKeyList;
+import nostr.event.list.TagList;
 import nostr.event.marshaller.impl.EventMarshaller;
 import nostr.event.marshaller.impl.TagListMarshaller;
 import nostr.event.marshaller.impl.TagMarshaller;
