+++ nostr-test/src/test/java/nostr/test/json/JsonTest.java
@@ -8,7 +8,6 @@ import nostr.util.UnsupportedNIPException;
 import nostr.event.impl.GenericEvent;
 import nostr.event.impl.TextNoteEvent;
 import nostr.event.marshaller.impl.EventMarshaller;
-import nostr.base.list.TagList;
 import nostr.event.marshaller.impl.TagMarshaller;
 import nostr.event.tag.DelegationTag;
 import nostr.event.tag.EventTag;
@@ -29,6 +28,7 @@ import java.util.logging.Level;
 import lombok.extern.java.Log;
 import nostr.base.GenericTagQuery;
 import nostr.event.impl.Filters;
+import nostr.event.list.TagList;
 import nostr.event.marshaller.impl.FiltersMarshaller;
 import nostr.event.marshaller.impl.GenericTagQueryMarshaller;
 import nostr.json.unmarshaller.impl.JsonExpressionUnmarshaller;
